feat: move rds client config to the client file
