+++ clients/client-rds-data/commands/ExecuteStatementCommand.ts
@@ -8,16 +8,17 @@ import {
   MiddlewareStack,
   SerdeContext
 } from "@aws-sdk/types";
-import { RdsDataServiceResolvedConfig } from "../RdsDtataServiceConfiguration";
 import { HttpRequest, HttpResponse } from "@aws-sdk/protocol-http";
 import {
   executeStatementAwsRestJson1_1Serialize,
   executeStatementAwsRestJson1_1Deserialize
 } from "../protocol/AwsRestJson1_1";
 import { ExecuteStatementRequest, ExecuteStatementResponse } from "../models";
-
-type InputTypesUnion = any;
-type OutputTypesUnion = any;
+import {
+  ServiceInputTypes,
+  ServiceOutputTypes,
+  RdsDataServiceResolvedConfig
+} from "../RdsDataServiceClient";
 
 export class ExecuteStatementCommand extends Command<
   ExecuteStatementRequest,
@@ -28,7 +29,7 @@ export class ExecuteStatementCommand extends Command<
   }
 
   resolveMiddleware(
-    clientStack: MiddlewareStack<InputTypesUnion, OutputTypesUnion>,
+    clientStack: MiddlewareStack<ServiceInputTypes, ServiceOutputTypes>,
     configuration: RdsDataServiceResolvedConfig,
     options?: HttpOptions
   ): Handler<ExecuteStatementRequest, ExecuteStatementResponse> {
